Implement unit tests for cleanup logic
