+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - [#852](https://github.com/spegel-org/spegel/pull/852) Remove use of Afero in Containerd config.
+- [#854](https://github.com/spegel-org/spegel/pull/854) Implement unit tests for cleanup logic .
 
 ### Deprecated
